protectData() uses a onStatusUpdate callback function + no more protectDataObservable
